+++ code/src/cljs/sixsq/nuvla/ui/deployment/utils.cljs
@@ -8,21 +8,18 @@
     [sixsq.nuvla.ui.utils.time :as time]
     [sixsq.nuvla.ui.utils.values :as values]))
 
+(def ^:const STARTED "STARTED")
+(def ^:const STARTING "STARTING")
+(def ^:const STOPPED "STOPPED")
+(def ^:const ERROR "ERROR")
 
-(def status-started "STARTED")
-(def status-starting "STARTING")
-(def status-stopped "STOPPED")
-(def status-error "ERROR")
-
-
-(defn status->icon
-  [status]
-  (let [icons-map {status-started  "fas fa-play"
-                   status-starting "fas fa-spinner"
-                   status-stopped  "fas fa-stop"
-                   status-error    "fas fa-exclamation"}]
-    (get icons-map status)))
-
+(defn state->icon
+  [state]
+  (let [icons-map {STARTED  "fas fa-play"
+                   STARTING "fas fa-spinner"
+                   STOPPED  "fas fa-stop"
+                   ERROR    "fas fa-exclamation"}]
+    (get icons-map state)))
 
 (defn resolve-url-pattern
   "When url-pattern is passed and all used params in pattern has values in
@@ -41,19 +38,16 @@
             (str/replace url param-pattern param-value))
           url-pattern pattern-value)))))
 
-
 (defn is-replicas-running?
   "Select all strings that end with 'replicas.running'."
   [[k _]]
   (and (string? k) (str/ends-with? k "replicas.running")))
-
 
 (defn positive-number?
   "Determines if the value is a positive number. Will not throw an exception
    if the argument is not a number."
   [n]
   (and (number? n) (pos? n)))
-
 
 (defn is-value-positive?
   [entry]
@@ -62,7 +56,6 @@
        :value
        general-utils/str->int
        positive-number?))
-
 
 (defn running-replicas?
   "Extracts the number of running replicas and returns true if the number is
@@ -75,37 +68,36 @@
          (every? true?))                                    ;; careful, this returns true for an empty collection!
     false))
 
-
 (defn deployment-in-transition?
   [state]
   (str/ends-with? (str state) "ING"))
-
 
 (defn assoc-delta-time
   "Given the start (as a string), this adds a :delta-time entry in minutes."
   [start {end :timestamp :as evt}]
   (assoc evt :delta-time (time/delta-minutes start end)))
 
-
-(defn is-started?
+(defn started?
   [state]
-  (= state "STARTED"))
+  (= state STARTED))
 
+(defn stopped?
+  [state]
+  (= state STOPPED))
 
 (defn deployment-href
   [id]
   (str "deployment/" (general-utils/id->uuid id)))
 
-
 (defn state-filter
   [state]
-  (case state
-    "STARTING" "state='RUNNING' or state='PENDING' or state='CREATED'"
+  (if (= state STARTING)
+    "state='RUNNING' or state='PENDING' or state='CREATED'"
     (str "state='" state "'")))
 
-
 (defn get-filter-param
-  [{:keys [full-text-search additional-filter state-selector nuvlabox module-id] :as _args}]
+  [{:keys [full-text-search additional-filter state-selector nuvlabox module-id]
+    :as   _args}]
   (let [filter-state     (when state-selector (state-filter state-selector))
         filter-nuvlabox  (when nuvlabox (str "nuvlabox='" nuvlabox "'"))
         filter-module-id (when module-id (str "module/id='" module-id "'"))
@@ -118,9 +110,9 @@
       full-text-search
       additional-filter)))
 
-
 (defn get-query-params
-  [{:keys [full-text-search additional-filter state-selector nuvlabox module-id page elements-per-page]}]
+  [{:keys [full-text-search additional-filter
+           state-selector nuvlabox module-id page elements-per-page]}]
   (let [filter-str (get-filter-param
                      {:full-text-search  full-text-search
                       :additional-filter additional-filter
@@ -133,7 +125,6 @@
      :orderby     "created:desc"
      :filter      filter-str}))
 
-
 (defn get-query-params-summary
   [full-text-search additional-filter]
   (let [full-text-search (general-utils/fulltext-query-string full-text-search)
@@ -145,11 +136,9 @@
              :last        0}
             (not (str/blank? filter-str)) (assoc :filter filter-str))))
 
-
 (defn is-selected?
   [selected-set id]
   (contains? selected-set id))
-
 
 (defn visible-deployment-ids
   [deployments]
@@ -158,16 +147,15 @@
        (map :id)
        set))
 
-
 (defn all-page-selected?
   [selected-set visible-deps-ids-set]
   (set/superset? selected-set visible-deps-ids-set))
 
-
 (defn format-nuvlabox-value
   [nuvlabox]
-  [:div [ui/Icon {:name "box"}] [values/as-link (general-utils/id->uuid nuvlabox) :page "edges"]])
-
+  [:div
+   [ui/Icon {:name "box"}]
+   [values/as-link (general-utils/id->uuid nuvlabox) :page "edges"]])
 
 (defn build-bulk-filter
   [{:keys [::spec/select-all?